requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Bcrypt==1.0.1
Flask-Mail==0.9.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
Pillow==10.1.0
requests==2.31.0
pytest==7.4.3
pytest-flask==1.3.0
python-decouple==3.8
gunicorn==21.2.0
Flask-Limiter==3.5.0

# Payment processing
stripe==7.7.0

# AWS S3 for file uploads
boto3==1.29.7

# Background tasks (optional)
celery==5.3.4
redis==5.0.1
